Split simple CLI commands
